kernel+userspace: Export kernel defs to userspace.
We can now selectively export kernel API/defs to userspace. This is so
libc can use the exact same datatypes/defines as the kernel. This can be
further extended to dxgmx specific functionality for example ioctls.

+++ drivers/fat/fat_read.c
@@ -11,7 +11,7 @@
 
 ssize_t fat_read(const VirtualNode* vnode, void* buf, size_t n, off_t off)
 {
-    if (n == 0 || off >= vnode->size)
+    if (n == 0 || (size_t)off >= vnode->size)
         return 0;
 
     FileSystem* fs = vnode->owner;
